ai-backend/requirements.txt
flask
flask-cors
flask-sse
anthropic
langchain
langchain-anthropic
langchain-community
chromadb
pymupdf
pdfplumber
python-dotenv
